increase entity size, fixes for ezby

+++ src/entity_type.zig
@@ -1,5 +1,5 @@
 /// Entity ID type
-pub const EntityId = u32;
+pub const EntityId = u64;
 
 /// An opaque representation of a entity which can be used to interact with it
 pub const Entity = packed struct {
